1.20 Update
Thanks daddy, gitingores are actually useless.....
so much for diff tracking....
also what the fuck is up with the zip files.....  they zip things an extra layer deep?  so what is even the point for a datapack for minecraft on github... might as fucking well ignore branches and just use zips in releases.  fuck me.